+++ src/service/collections/undelete.rs
@@ -97,7 +97,7 @@ fn undelete_document_pool(ctx: &ServiceContext, id: &str) -> Result<Document> {
     let pool = ctx.pool.context("pool required")?;
     let runner = ctx.runner()?;
     let def = ctx.collection_def()?;
-    let mut conn = pool.get().context("DB connection")?;
+    let mut conn = pool.write().context("DB connection")?;
     let tx = conn.transaction_immediate().context("Start transaction")?;
 
     let queue = Rc::new(RefCell::new(Vec::new()));