+++ src/service/versions/restore.rs
@@ -155,7 +155,7 @@ fn restore_collection_version_pool(
     let pool = ctx.pool.context("pool required")?;
     let runner = ctx.runner()?;
     let def = ctx.collection_def()?;
-    let mut conn = pool.get().context("DB connection")?;
+    let mut conn = pool.write().context("DB connection")?;
     let tx = conn.transaction_immediate().context("Start transaction")?;
 
     let infra = LuaCrudInfra::from_ctx(ctx, None, None);
@@ -340,7 +340,7 @@ pub fn restore_global_version(
     let pool = ctx.pool.context("pool required")?;
     let runner = ctx.runner()?;
     let def = ctx.global_def()?;
-    let mut conn = pool.get().context("DB connection")?;
+    let mut conn = pool.write().context("DB connection")?;
     let tx = conn.transaction_immediate().context("Start transaction")?;
 
     let infra = LuaCrudInfra::from_ctx(ctx, None, None);